backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
fastapi==0.120.2
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
idna==3.11
pydantic==2.12.3
pydantic_core==2.41.4
smmap==5.0.2
sniffio==1.3.1
starlette==0.49.1
tree-sitter==0.25.2
tree-sitter-python>=0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
